requirements.txt
Babel==2.11.0
folium==0.18.0
inflection==0.5.1
numpy==2.2.0
pandas==2.2.3
Pillow==11.0.0
plotly==5.22.0
streamlit==1.32.0
streamlit_folium==0.23.2